requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
passlib[bcrypt]
pyjwt
python-jose[cryptography]
email-validator
pydantic[email]